reorg; begin config implementation for shipswitcher
