+++ config.ini
@@ -17,7 +17,7 @@ Fleet4: 38
 Enabled: True
 
 [Combat]
-Enabled: False
+Enabled: True
 Engine: legacy
 FleetMode:
 Map: 1-1
@@ -33,6 +33,21 @@ LBASGroup1Nodes:
 LBASGroup2Nodes:
 LBASGroup3Nodes:
 MiscOptions:
+
+[ShipSwitcher]
+Enabled: True
+Slot1Criteria: damage
+Slot1Ships: N:E:20, N:S:32
+Slot2Criteria: sparkle
+Slot2Ships: C:SS:>20:L:R
+Slot3Criteria:
+Slot3Ships:
+Slot4Criteria:
+Slot4Ships:
+Slot5Criteria:
+Slot5Ships:
+Slot6Criteria:
+Slot6Ships:
 
 [Quests]
 Enabled: True
